Campaign exceptional handling

+++ src/main/java/sg/edu/nus/iss/springboot/voucher/management/service/impl/CampaignService.java
@@ -20,7 +20,7 @@ import sg.edu.nus.iss.springboot.voucher.management.service.ICampaignService;
 import sg.edu.nus.iss.springboot.voucher.management.utility.DTOMapper;
 
 @Service
-public class CampaignService implements ICampaignService        {
+public class CampaignService implements ICampaignService {
 
     private static final Logger logger = LoggerFactory.getLogger(CampaignService.class);
 
@@ -35,9 +35,12 @@ public class CampaignService implements ICampaignService        {
 
     @Override
     public List<CampaignDTO> findAllCampaigns() {
+
+        logger.info("Getting all campaigns...");
         List<Campaign> campaigns = campaignRepository.findAll();
+        logger.info("Found {}, converting to DTOs...", campaigns.size());
         List<CampaignDTO> campaignDTOs = new ArrayList<CampaignDTO>();
-        for (Campaign campaign: campaigns){
+        for (Campaign campaign : campaigns) {
             campaignDTOs.add(DTOMapper.toCampaignDTO(campaign));
         }
         return campaignDTOs;
@@ -45,41 +48,67 @@ public class CampaignService implements ICampaignService        {
 
     @Override
     public CampaignDTO findByCampaignId(String campaignId) {
+        logger.info("Getting campaign for campaignId {} ...", campaignId);
         Campaign campaign = campaignRepository.findById(campaignId).orElse(null);
-        if (campaign != null)
-        {
+        if (campaign != null) {
+            logger.info("Campaign found...");
             return DTOMapper.toCampaignDTO(campaign);
         }
+        logger.warn("Didn't find any campaign for campaignId {}...", campaignId);
         return null;
     }
 
     @Override
     public CampaignDTO create(Campaign campaign) {
-        User user = userRepository.findByEmail(campaign.getCreatedBy().getEmail());
-        Store store = storeRepository.findById(campaign.getStore().getStoreId()).orElseThrow();
-        campaign.setCreatedBy(user);
-        campaign.setCreatedDate(LocalDateTime.now());
-        campaign.setUpdatedBy(user);
-        campaign.setUpdatedDate(LocalDateTime.now());
-        campaign.setStore(store);
-        Campaign savedCampaign = campaignRepository.save(campaign);
-        CampaignDTO campaignDTO = DTOMapper.toCampaignDTO(savedCampaign);
-        return campaignDTO;
+        try {
+            User user = userRepository.findByEmail(campaign.getCreatedBy().getEmail());
+            Store store = storeRepository.findById(campaign.getStore().getStoreId()).orElseThrow();
+            campaign.setCreatedBy(user);
+            campaign.setCreatedDate(LocalDateTime.now());
+            campaign.setUpdatedBy(user);
+            campaign.setUpdatedDate(LocalDateTime.now());
+            campaign.setStore(store);
+            logger.info("Saving campaign...");
+            Campaign savedCampaign = campaignRepository.save(campaign);
+            logger.info("Saved successfully...");
+            return DTOMapper.toCampaignDTO(savedCampaign);
+        } catch (Exception ex) {
+            logger.error("Campaign saving exception... {}", ex.toString());
+            return null;
+        }
+
     }
 
     @Override
     public CampaignDTO update(Campaign campaign) {
-        User user = userRepository.findByEmail(campaign.getCreatedBy().getEmail());
-        campaign.setUpdatedBy(user);
-        campaign.setUpdatedDate(LocalDateTime.now());
-        Campaign savedCampaign = campaignRepository.save(campaign);
-        CampaignDTO campaignDTO = DTOMapper.toCampaignDTO(savedCampaign);
-        return campaignDTO;
+        try {
+            User user = userRepository.findByEmail(campaign.getCreatedBy().getEmail());
+            campaign.setUpdatedBy(user);
+            campaign.setUpdatedDate(LocalDateTime.now());
+            logger.info("Saving campaign...");
+            Campaign savedCampaign = campaignRepository.save(campaign);
+            logger.info("Saved successfully...");
+            CampaignDTO campaignDTO = DTOMapper.toCampaignDTO(savedCampaign);
+            return campaignDTO;
+        } catch (Exception ex) {
+            logger.error("Campaign updating exception... {}", ex.toString());
+            return null;
+        }
+
     }
 
     @Override
-	public void delete(String campaignId) {
-        campaignRepository.findById(campaignId).ifPresent(campaign -> campaignRepository.delete(campaign));
-	}
-
+    public void delete(String campaignId) {
+        if (campaignId == null) {
+            logger.info("CampaignId is null...");
+            return;
+        }
+        try {
+            logger.info("Deleting campaignId {}...", campaignId);
+            campaignRepository.findById(campaignId).ifPresent(campaign -> campaignRepository.delete(campaign));
+            logger.info("Deleted successfully...");
+        } catch (Exception ex) {
+            logger.error("Campaign deleting exception... {}", ex.toString());
+        }
+    }
 }
